@@ -6,51 +6,51 @@
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "start:remote": "ng serve --configuration development-remote",
-    "build:remote": "ng build --configuration production-remote"    
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.1.0",
-    "@angular/cdk": "^18.1.0",
-    "@angular/common": "^18.1.0",
-    "@angular/compiler": "^18.1.0",
-    "@angular/core": "^18.1.0",
-    "@angular/forms": "^18.1.0",
-    "@angular/material": "^18.1.0",
-    "@angular/platform-browser": "^18.1.0",
-    "@angular/platform-browser-dynamic": "^18.1.0",
-    "@angular/router": "^18.1.0",
+    "@angular/animations": "^18.2.0",
+    "@angular/common": "^18.2.0",
+    "@angular/compiler": "^18.2.0",
+    "@angular/core": "^18.2.0",
+    "@angular/forms": "^18.2.0",
+    "@angular/platform-browser": "^18.2.0",
+    "@angular/platform-browser-dynamic": "^18.2.0",
+    "@angular/router": "^18.2.0",
     "@microsoft/signalr": "^8.0.7",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.1",
+    "@popperjs/core": "^2.11.8",
     "angular-i18next": "^18.0.0",
     "bcryptjs": "^2.4.3",
-    "i18next": "^23.12.1",
+    "bootstrap": "^5.3.2",
+    "i18next": "^23.16.1",
     "i18next-browser-languagedetector": "^8.0.0",
-    "i18next-http-backend": "^2.5.2",
-    "ngx-indexed-db": "^19.0.0",
+    "i18next-http-backend": "^2.6.2",
+    "ngx-indexed-db": "^19.0.1",
     "ngx-logger": "^5.0.12",
-    "ngx-progressbar": "^11.1.0",
-    "ngx-store": "^3.1.1",
+    "ngx-progressbar": "^12.0.2",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "ua-parser-js": "^1.0.38",
-    "zone.js": "~0.14.3"
+    "ua-parser-js": "^1.0.39",
+    "uuid": "^10.0.0",
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.1.0",
-    "@angular/cli": "^18.1.0",
-    "@angular/compiler-cli": "^18.1.0",
+    "@angular-devkit/build-angular": "^18.2.9",
+    "@angular/cli": "^18.2.9",
+    "@angular/compiler-cli": "^18.2.0",
+    "@angular/localize": "^18.2.0",
     "@types/bcryptjs": "^2.4.6",
     "@types/jasmine": "~5.1.0",
     "@types/ua-parser-js": "^0.7.39",
-    "jasmine-core": "~5.1.0",
+    "@types/uuid": "^10.0.0",
+    "jasmine-core": "~5.2.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "prettier": "^3.3.3",
     "typescript": "~5.5.2"
   }
 }
