@@ -1,40 +1,57 @@
-import { CommonModule } from '@angular/common';
-import { Component, Renderer2 } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
-import { NgProgressComponent } from 'ngx-progressbar';
-import { Subscription } from 'rxjs';
-import { NavigationComponent } from './navigation/navigation.component';
+import { Component, computed, effect, inject, Renderer2, signal } from '@angular/core';
+import { Router, RouterLink, RouterLinkActive, RouterOutlet } from '@angular/router';
+import { I18NextModule } from 'angular-i18next';
+import { NgProgressbar } from 'ngx-progressbar';
+import { NgProgressHttp } from 'ngx-progressbar/http';
+import { SvgComponent } from './components/svg/svg.component';
+import { AuthenticationService } from './services/authentication.service';
 import { ThemeService } from './services/theme.service';
+import { ToastsComponent } from './components/toast/toast.component';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, NavigationComponent, NgProgressComponent],
-  providers: [CommonModule, ThemeService],
+  imports: [RouterOutlet, NgProgressbar, NgProgressHttp, SvgComponent, I18NextModule, RouterLink, RouterLinkActive, ToastsComponent],
   templateUrl: './app.component.html',
+  styleUrl: './app.component.scss',
 })
 export class AppComponent {
-  themeSubscription?: Subscription;
+  private router = inject(Router);
+  private renderer = inject(Renderer2);
+  private themeService = inject(ThemeService);
+  private authenticationService = inject(AuthenticationService);
 
-  constructor(
-    private themeService: ThemeService,
-    private renderer: Renderer2
-  ) {}
+  isLoggedIn = computed(() => !!this.authenticationService.activeUser());
 
-  get primaryColor() {
-    return this.themeService.primaryColor;
-  }
+  constructor() {
+    effect(() => {
+      const theme = this.themeService.theme();
+      let scheme = this.themeService.scheme();
 
-  ngOnInit() {
-    this.themeSubscription = this.themeService.theme$.subscribe((theme) => {
-      if (theme.previous) {
-        this.renderer.removeClass(document.body, theme.previous);
+      if (scheme === 'auto') {
+        scheme = window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
       }
-      this.renderer.addClass(document.body, theme.current);
+      this.renderer.setAttribute(window.document.body, 'data-bs-theme', scheme);
+
+      if (theme) {
+        this.renderer.removeClass(window.document.body, theme);
+      }
     });
   }
 
-  ngOnDestroy() {
-    this.themeSubscription?.unsubscribe();
+  setLight(): void {
+    this.themeService.scheme.set('light');
+  }
+
+  setDark(): void {
+    this.themeService.scheme.set('dark');
+  }
+
+  setAuto(): void {
+    this.themeService.scheme.set('auto');
+  }
+
+  logout(): void {
+    this.authenticationService.logout().subscribe(() => this.router.navigate(['/login']));
   }
 }
